+++ admin/hapus_pgna.php
@@ -6,7 +6,7 @@ require "helpers/auth.php";
 if (isset($_GET["id"])) {
     $id = $_GET["id"];
     try {
-        query("DELETE FROM pengguna WHERE id = :id", [':id' => $id]);
+        query("DELETE FROM pengguna WHERE role != 'admin' AND id = :id", [':id' => $id]);
         setAlert('success', 'Pengguna berhasil dihapus');
         redirect('daftar_pgna.php');
     } catch (Exception $error) {
